accessible/tests/mochitest/actions/test_inputs.html

<html>

<head>
  <title>nsIAccessible actions testing for inputs</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../actions.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      var actionsArray = [
        {
          ID: "button",
          actionName: "press",
          events: CLICK_EVENTS
        },
        {
          ID: "checkbox_unchecked",
          actionName: "check",
          events: CLICK_EVENTS
        },
        {
          ID: "checkbox_checked",
          actionName: "uncheck",
          events: CLICK_EVENTS
        },
        {
          ID: "checkbox_mixed",
          actionName: "cycle",
          events: CLICK_EVENTS
        },
        {
          ID: "radio",
          actionName: "select",
          events: CLICK_EVENTS
        }
      ];
      document.getElementById("checkbox_mixed").indeterminate = true;

      testActions(actionsArray);
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=477975"
     title="nsIAccessible actions testing">
    Mozilla Bug 477975
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <button id="button" value="button">Button</button>

  <input id="checkbox_unchecked" type="checkbox">Checkbox</input>

  <input id="checkbox_checked" type="checkbox" checked="true">Checkbox</input>

  <input id="checkbox_mixed" type="checkbox">Checkbox</input>

  <fieldset>
    <input id="radio" type="radio">Radio</input>
  </fieldset>
</body>
</html>
